Implemented procedures listing
Right now new procedures are defaulted with 'Untitled' for title and
'Nameless' for author fields. These should either be nullable on the
server. Alternatively, author should default as the username and title
should default to 'Untitled' similar to Google Docs.

It also doesn't redirect the user to the main editor for the procedure
after creating a new procedure. This will be implemented in #216

Other Changes:
- Removed trailing slashes in API resource endpoints (e.g.
  /{resource_name}/{id} instead of ../{id}/
- Renamed templates to be more specific (e.g. auth -> auth_layout)
- Prepared for i18n

Resolves #215, Closes #167